--- layout/index.html
@@ -1,88 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link
-      rel="stylesheet"
-      href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
-      integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm"
-      crossorigin="anonymous"
-    />
-    <title>Week one api</title>
-  </head>
-
-  <body class="container">
-    <br />
-    <div class="d-flex justify-content-center justify-content-md-center">
-      <div class="col-md-6">
-        <form class="form-inline" id="searchForm">
-          <div class="form-group">
-            <label for="inputText">Enter Type of Food: </label>
-            <input
-              type="text"
-              name="q"
-              id="food"
-              class="form-control mx-sm-3"
-              aria-describedby="passwordHelpInline"
-              placeholder="E.g: Chicken"
-              value=""
-            />
-          </div>
-
-          <div class="form-group">
-            <label for="inputText">Enter Health notice: </label>
-            <input
-              type="text"
-              id="health"
-              name="health"
-              class="form-control mx-sm-3"
-              aria-describedby="passwordHelpInline"
-              placeholder="peanut-free"
-              value=""
-            />
-          </div>
-
-          <div class="form-group">
-            <label for="inputText">Enter Amount of Calories: </label>
-            <input
-              type="text"
-              id="calories"
-              name="calories"
-              class="form-control mx-sm-3"
-              aria-describedby="passwordHelpInline"
-              placeholder="E.g: 500"
-              value=""
-              autocomplete="off"
-            />
-          </div>
-
-          <button type="submit" id="submit" class="btn btn-outline-primary">
-            Check for Recipes
-          </button>
-        </form>
-        <br /><br />
-        <a href="user" class="btn btn-outline-primary">Go to login page</a>
-        <br />
-        <div id="q"></div>
-      </div>
-    </div>
-  </body>
-  <script
-    src="https://code.jquery.com/jquery-3.4.1.slim.min.js"
-    integrity="sha384-J6qa4849blE2+poT4WnyKhv5vZF5SrPo0iEjwBvKU7imGFAV0wwj1yYfoRSJoZ+n"
-    crossorigin="anonymous"
-  ></script>
-  <script
-    src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js"
-    integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo"
-    crossorigin="anonymous"
-  ></script>
-  <script
-    src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js"
-    integrity="sha384-wfSDF2E50Y2D1uUdj0O3uMBJnjuUD4Ih7YwaYd1iqfktj0Uod8GCExl3Og8ifwB6"
-    crossorigin="anonymous"
-  ></script>
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
-  <script type="text/javascript" src="/Src/index.js"></script>
-</html>
