--- layout/login.html
@@ -1,52 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="login-styles.css">
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
-        integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-    <title>Login</title>
-</head>
-
-<body class="container">
-    <img src="../layout/images/background-front-page.jpeg" id="bg" alt="">
-    <div class="d-flex justify-content-center">
-        <!-- image and login form -->
-        <div class="image">
-            <img style="height: 500px; width: 500px; margin-left: 20%;" src="../layout/images/front-page-image.png"
-                alt="Front page image">
-        </div>
-
-        <div class="login">
-            <h1>App_Name</h1>
-            <div class="d-flex justify-content-center">
-                <form method="post" action="">
-                    <p><input type="text" name="login" value="" placeholder="Username or Email"></p>
-                    <p><input type="password" name="password" value="" placeholder="Password"></p>
-
-                    <br />
-
-                    <div class="d-flex justify-content-center">
-                        <button type="submit" id="submit" class="btn btn-outline-primary ">
-                            Login
-                        </button>
-                    </div>
-            </div>
-            <br />
-            <hr>
-
-            <div class="login-help">
-                <p>Forgot your password? <a href="#">Click here to reset it</a>.</p>
-
-                <hr />
-
-                <p>Don't have an account? <a href="../layout/signup.html">Click here to sign up</a>.</p>
-            </div>
-            </form>
-        </div>
-    </div>
-</body>
-
-</html>
