--- layout/signup.html
@@ -1,57 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="login-styles.css">
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
-        integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-    <title>Sign up</title>
-</head>
-<body class="container">
-    <img src="../layout/images/background-front.jpeg" id="bg" alt="">
-
-    <div class="login">
-        <h1 style="background-color: #e8e8e8">App_Name</h1>
-        <div class="d-flex justify-content-center">
-            <form method="post" action="./login.html">
-                <p><input type="text" name="login" value="" placeholder="Username"></p>
-                <p><input type="text" name="login" value="" placeholder="Email"></p>
-                <p><input type="text" name="login" value="" placeholder="Full Name"></p>
-                <p><input type="password" name="password" value="" placeholder="Password"></p>
-
-                <br />
-                <hr />
-                <div class="d-flex justify-content-center">
-                    <button onclick="registered()" type="submit" id="submit" class="btn btn-outline-success">
-                        Sign up
-                    </button>
-
-                    <script>
-                        let registered = () => {
-                            alert("You have been successfully registered!")
-                        }
-                    </script>
-                </div>
-        </div>
-        <br />
-        </form>
-    </div>
-
-    <div class="login">
-    
-        <div class="login-help">
-            <hr />
-            <p>Already have an account? <a style="color: green;" href="../layout/login.html">Click here to login</a>.</p>
-
-            <hr />
-        </div>
-                
-        </div>
-        <br />
-        </form>
-    </div>
-</div>
-    
-</body>
-</html>
