+++ views/login.ejs
@@ -3,123 +3,80 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="css/login-styles.css" />
     <link
       rel="stylesheet"
-      href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
-      integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm"
+      href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
+      integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T"
       crossorigin="anonymous"
     />
     <title>Login</title>
   </head>
 
-  <body>
-    <div class="container">
-      <h1>Login</h1>
-      <form
-        class="form-group"
-        id="login-form"
-        method="POST"
-        action="/user/gNQu5jGgxPL42r8g5zm6"
-      >
-        <!-- <form class="form-group" id="login-form"> -->
-        <div class="form-group">
-          <label for="username">Username: </label>
-          <input
-            type="text"
-            name="loginUser"
-            id="username"
-            class="form-control mx-sm-3"
-            value=""
-            required
-          />
-        </div>
+  <body class="container">
+    <img src="images/background-front.jpeg" id="bg" alt="" />
+    <div class="d-flex justify-content-center">
+      <!-- image and login form -->
 
-        <div class="form-group">
-          <label for="password">password: </label>
-          <input
-            type="password"
-            name="loginPw"
-            id="password"
-            class="form-control mx-sm-3"
-            value=""
-            required
-          />
-        </div>
-
-        <button type="submit" id="login-submit" class="btn btn-outline-primary">
-          Login
-        </button>
-      </form>
-    </div>
-
-    <div class="container">
-      <h1>Sign-up</h1>
-      <form
-        class="form-group"
-        id="sign-up-form"
-        method="POST"
-        action="/user/JKp7DeJXgaFtxaJ7FTXb"
-      >
-        <!-- <form class="form-group" id="sign-up-form"> -->
-        <div class="form-group">
-          <label for="sign-up-username">Username: </label>
-          <input
-            type="text"
-            name="signupUser"
-            id="sign-up-username"
-            class="form-control mx-sm-3"
-            value=""
-            required
-          />
-        </div>
-
-        <div class="form-group">
-          <label for="sign-up-password">password: </label>
-          <input
-            type="password"
-            name="signupPw"
-            id="sign-up-password"
-            class="form-control mx-sm-3"
-            value=""
-            required
-          />
-        </div>
-
-        <button
-          type="submit"
-          id="sign-up-submit"
-          class="btn btn-outline-primary"
+      <div class="login">
+        <h1
+          style="background-color: #71B340; color: #65636C; font-size:x-large;"
         >
-          Sign-up
-        </button>
-      </form>
+          App_Name
+        </h1>
+        <div class="d-flex justify-content-center">
+          <form
+            id="login-form"
+            method="post"
+            action="/user/gNQu5jGgxPL42r8g5zm6"
+          >
+            <div class="form-group">
+              <input
+                type="text"
+                name="loginUser"
+                id="username"
+                class="form-control mx-sm-3"
+                value=""
+                placeholder="Email or username"
+                required
+              />
+            </div>
+
+            <div class="form-group">
+              <input
+                type="password"
+                name="loginPw"
+                id="password"
+                class="form-control mx-sm-3"
+                value=""
+                placeholder="Password"
+                required
+              />
+            </div>
+            <div class="d-flex justify-content-center">
+              <button type="submit" id="submit" class="btn btn-outline-success">
+                Login
+              </button>
+            </div>
+            <hr />
+
+            <div class="login-help">
+              <p>
+                Forgot your password?
+                <a style="color: green;" href="#">Click here to reset it</a>.
+              </p>
+
+              <hr />
+
+              <p>
+                Don't have an account?
+                <a style="color: green;" href="/signup">Click here to sign up</a
+                >.
+              </p>
+            </div>
+          </form>
+        </div>
+      </div>
     </div>
-
-    <!-- <%- include('./parts/jq'); -%>
-  <script>
-    $(function () {
-      $("#login-form").on('submit', function () {
-        $.ajax({
-          url: "/login/gNQu5jGgxPL42r8g5zm6",
-          type: "post",
-          data: {
-            loginUser: $("#username").val(),
-            loginPw: $("#password").val(),
-          }
-        })
-      });
-
-      $("#sign-up-form").on('submit', function () {
-        $.ajax({
-          url: "/login/JKp7DeJXgaFtxaJ7FTXb",
-          type: "post",
-          data: {
-            username: $("#sign-up-username").val(),
-            password: $("#sign-up-password").val(),
-          }
-        })
-      });
-    });
-  </script> -->
   </body>
 </html>
